+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.5.1 (2023-06-22)
 
 - Regenerate graph client for new module library APIs.
+- Add streaming API to download blobs.
 
 
 # 0.5.0 (2023-04-18)